[Rust][FFI] Split lib.rs into per-surface modules
Split the ~2,250-line ffi/src/lib.rs into per-surface modules (common,
arrow, builder, sdk, stream, proto_schema); lib.rs now holds only module
declarations and re-exports. Pure refactor: no API, ABI, or behavior
change, and zerobus.h is byte-identical. Module declaration order is
load-bearing for cbindgen's output order; build.rs now watches the whole
src/ tree so the header regenerates on any module change.
