+++ rust/ffi/NEXT_CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Internal Changes
 
+- Split `src/lib.rs` into per-surface modules (`common`, `arrow`, `builder`, `sdk`, `stream`, `proto_schema`); `lib.rs` now holds only module declarations and re-exports. Pure refactor — no API, ABI, or behavior change; `zerobus.h` is byte-identical. `build.rs` now watches the whole `src/` tree so the header regenerates on any module change.
+
 ### Behavior Changes
 
 ### Breaking Changes
